Add support to fix EC signatures in DER format from external signers. Check for DER EC signatures during validation.
