+++ sdk/src/cose_sign.rs
@@ -34,6 +34,7 @@ use crate::{
         cose_timestamp_countersign, cose_timestamp_countersign_async, make_cose_timestamp,
     },
     trust_handler::TrustHandlerConfig,
+    utils::sig_utils::{der_to_p1363, parse_ec_der_sig},
     AsyncSigner, Error, Result, Signer, SigningAlg,
 };
 
@@ -176,11 +177,23 @@ pub(crate) fn cose_sign(signer: &dyn Signer, data: &[u8], box_size: usize) -> Re
         sign1.payload.as_ref().unwrap_or(&vec![]),
     );
 
-    if _sync {
-        sign1.signature = signer.sign(&tbs)?;
+    let signature = if _sync {
+        signer.sign(&tbs)?
     } else {
-        sign1.signature = signer.sign(tbs).await?;
-    }
+        signer.sign(tbs).await?
+    };
+
+    sign1.signature = match alg {
+        SigningAlg::Es256 | SigningAlg::Es384 | SigningAlg::Es512 => {
+            if parse_ec_der_sig(&signature).is_ok() {
+                // fix up DER signature to be in P1363 format
+                der_to_p1363(&signature, alg)?
+            } else {
+                signature
+            }
+        }
+        _ => signature,
+    };
 
     sign1.payload = None; // clear the payload since it is known
